# backend/docs/swagger.yaml
definitions:
...
  internal_handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - password
    type: object
...
  internal_handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      nickname:
        maxLength: 50
        minLength: 1
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - nickname
    - password
    type: object
  internal_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with phone/email and password, returns JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: User login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Use a refresh token to get a new access/refresh token pair
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register with phone/email, password, and nickname
      parameters:
      - description: Registration info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"